+++ CHANGES.md
@@ -4,3 +4,4 @@
 * changed dependency versions
 ## 0.2.4
 * changing deprecated `std::env::home_dir` to `dirs::home_dir`
+* improved the efficiency of --strict and --prove